add settings
These settings are directly taken from the old version of the repository
which was managed by pathogen. They are not correctly ordered yet and
some things might appear in wrong order or file.
The old 'bootstrap' and 'config' folders were merged into this one
'config' folder because Vundle does not seem to need any bootstrapping
for the dirsettings plugin to work properly.